requirements.txt
Appium-Python-Client
selenium
python-telegram-bot